requirements.txt
pyyaml
numpy
rich
wandb
tensorboard
torchtyping